+++ crates/eval/Cargo.toml
@@ -9,12 +9,14 @@ agent.workspace = true
 anyhow.workspace = true
 assistant_tool.workspace = true
 assistant_tools.workspace = true
+assistant_settings.workspace = true
 client.workspace = true
 collections.workspace = true
 context_server.workspace = true
 dap.workspace = true
 env_logger.workspace = true
 fs.workspace = true
+futures.workspace = true
 gpui.workspace = true
 gpui_tokio.workspace = true
 language.workspace = true
